# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SETU-DRR API",
    "description": "Hazard Red Zone Identification, Carrying Capacity Assessment & Relocation Decision Support Platform.",
    "version": "v1.0.0"
  },
  "paths": {
...
    "/habitations/{id}/sites": {
      "get": {
        "tags": [
          "Habitations & Triage"
        ],
        "summary": "Get ranked candidate relocation sites for a habitation",
        "description": "Retrieves candidate relocation sites within search radius (default 15 km) ranked by suitability, carrying capacity, and distance. Includes full resource capacity breakdowns and binding constraints.",
        "operationId": "get_habitation_candidate_sites_habitations__id__sites_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Habitation ID (integer primary key).",
              "examples": [
                1
              ],
              "title": "Id"
            },
            "description": "Habitation ID (integer primary key)."
          },
          {
            "name": "radius_km",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 100.0,
                  "exclusiveMinimum": 0.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Configurable search radius around habitation in kilometers (default 15 km).",
              "title": "Radius Km"
            },
            "description": "Configurable search radius around habitation in kilometers (default 15 km)."
          },
          {
            "name": "min_suitability",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by minimum composite suitability score (0-100).",
              "title": "Min Suitability"
            },
            "description": "Filter by minimum composite suitability score (0-100)."
          },
          {
            "name": "include_screening",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Whether to include unverified screening-grade candidate sites (default False).",
              "default": false,
              "title": "Include Screening"
            },
            "description": "Whether to include unverified screening-grade candidate sites (default False)."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Number of records per page (max 200).",
              "default": 50,
              "title": "Limit"
            },
            "description": "Number of records per page (max 200)."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Pagination offset.",
              "default": 0,
              "title": "Offset"
            },
            "description": "Pagination offset."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_CandidateSiteItem_"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Requested resource, cell, or entity does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error - Request parameter or payload validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An unexpected system or database error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable - No valid serving version is active. Pipeline data is not ready.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/habitations/{id}/external-recommendations": {
      "get": {
        "tags": [
          "External Recommendations & Benchmark"
        ],
        "summary": "Get external pipeline relocation recommendations for a habitation",
        "description": "Retrieves offline GIS candidate proposals for this habitation from partner pipelines. These are external evidence records and do not represent authoritative SETU allocation decisions.",
        "operationId": "get_habitation_external_recommendations_habitations__id__external_recommendations_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Habitation ID (integer primary key).",
              "examples": [
                1
              ],
              "title": "Id"
            },
            "description": "Habitation ID (integer primary key)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExternalRecommendationItem"
                  },
                  "title": "Response Get Habitation External Recommendations Habitations  Id  External Recommendations Get"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Requested resource, cell, or entity does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error - Request parameter or payload validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An unexpected system or database error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable - No valid serving version is active. Pipeline data is not ready.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AugmentedCapacityDTO": {
        "properties": {
          "relieved_constraint": {
            "$ref": "#/components/schemas/BindingConstraint"
          },
          "augmented_capacity": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Augmented Capacity",
            "description": "Augmented carrying capacity in households."
          },
          "next_binding_constraint": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BindingConstraint"
              },
              {
                "type": "null"
              }
            ],
            "description": "The secondary resource constraint that limits the site after relief."
          },
          "indicative_intervention": {
            "type": "string",
            "title": "Indicative Intervention",
            "description": "Recommended engineering or policy intervention."
          },
          "indicative_cost_inr_lakhs": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Indicative Cost Inr Lakhs",
            "description": "Estimated intervention cost in Lakhs INR (None if unverified)."
          }
        },
        "type": "object",
        "required": [
          "relieved_constraint",
          "augmented_capacity",
          "indicative_intervention"
        ],
        "title": "AugmentedCapacityDTO",
        "description": "Augmentation relief assessment for relieving the primary binding constraint."
      },
      "BindingConstraint": {
        "type": "string",
        "enum": [
          "land",
          "water",
          "school",
          "health"
        ],
        "title": "BindingConstraint",
        "description": "Binding capacity constraints for destination sites (PRD \u00a76.8)."
      },
...
      "CandidateSiteItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "distance_km": {
            "type": "number",
            "minimum": 0.0,
            "title": "Distance Km",
            "description": "Geodesic distance from source habitation in km."
          },
          "area_ha": {
            "type": "number",
            "minimum": 0.0,
            "title": "Area Ha",
            "description": "Total contiguous developable area in hectares."
          },
          "tenure": {
            "$ref": "#/components/schemas/TenureType",
            "description": "Tenure status (government_revenue, private, tenure_unverified)."
          },
          "slope_mean": {
            "type": "number",
            "title": "Slope Mean",
            "description": "Mean terrain slope in degrees.",
            "default": 0.0
          },
          "mhi_max": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Mhi Max",
            "description": "Maximum static multi-hazard index inside site (None if unmeasured)."
          },
          "suitability": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Suitability",
            "description": "Composite suitability score (0-100, None if unassigned/provisional), separate from capacity."
          },
          "capacity": {
            "$ref": "#/components/schemas/CapacityBreakdownDTO",
            "description": "Full resource capacity breakdown."
          },
          "augmented": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AugmentedCapacityDTO"
              },
              {
                "type": "null"
              }
            ],
            "description": "Capacity outcome if binding bottleneck is relieved."
          },
          "centroid": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Centroid",
            "description": "[longitude, latitude] coordinates."
          },
          "assessment_status": {
            "type": "string",
            "title": "Assessment Status",
            "description": "Assessment status: 'fully_assessed', 'partial', or 'screening_only'.",
            "default": "screening_only"
          },
          "eligibility_status": {
            "type": "string",
            "title": "Eligibility Status",
            "description": "Eligibility status: 'eligible', 'ineligible', or 'unknown'.",
            "default": "unknown"
          },
          "allocatable": {
            "type": "boolean",
            "title": "Allocatable",
            "description": "True only if site is fully eligible under canonical CandidateSitePolicy.",
            "default": false
          },
          "rejection_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rejection Reasons",
            "description": "Reasons why site is not currently allocatable."
          },
          "source_site_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Site Id",
            "description": "External source record identifier if imported."
          },
          "screening_grade": {
            "type": "string",
            "title": "Screening Grade",
            "description": "Persistent decision-support disclaimer notice.",
            "default": "Screening Grade: Cell-level screening and prioritisation tool. Geotechnical investigation, hydraulic study, and community consultation required before executing relocation orders."
          }
        },
        "type": "object",
        "required": [
          "id",
          "distance_km",
          "area_ha",
          "tenure",
          "capacity",
          "centroid"
        ],
        "title": "CandidateSiteItem",
        "description": "Candidate site card in ranked destination list (GET /habitations/{id}/sites)."
      },
      "CapacityBreakdownDTO": {
        "properties": {
          "cc_land": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Cc Land",
            "description": "Households supportable by developable land area."
          },
          "land_screening_capacity": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Land Screening Capacity",
            "description": "Households supportable by developable land area (screening-only capacity).",
            "default": 0
          },
          "cc_water": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Cc Water",
            "description": "Households supportable by sustainable potable water yield (None if unmeasured)."
          },
          "cc_school": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Cc School",
            "description": "Households supportable by spare school capacity (None if unmeasured)."
          },
          "cc_health": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Cc Health",
            "description": "Households supportable by spare primary health capacity (None if unmeasured)."
          },
          "livelihood_multiplier": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Livelihood Multiplier",
            "description": "Multiplier for economic connectivity."
          },
          "cc_final": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Cc Final",
            "description": "Binding minimum carrying capacity in households (None if lifelines unassessed)."
          },
          "binding_constraint": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BindingConstraint"
              },
              {
                "type": "null"
              }
            ],
            "description": "The primary limiting capacity bottleneck (None if lifelines unassessed)."
          },
          "tied_constraints": {
            "items": {
              "$ref": "#/components/schemas/BindingConstraint"
            },
            "type": "array",
            "title": "Tied Constraints",
            "description": "All resource constraints matching the minimum bottleneck value."
          },
          "assessment_status": {
            "type": "string",
            "title": "Assessment Status",
            "description": "Assessment completeness: 'fully_assessed', 'partial', or 'screening_only'.",
            "default": "screening_only"
          },
          "data_quality": {
            "type": "string",
            "title": "Data Quality",
            "description": "Data quality state: 'complete', 'partial', or 'unavailable'.",
            "default": "complete"
          },
          "policy_version": {
            "type": "string",
            "title": "Policy Version",
            "description": "Audit version of the capacity policy norms applied.",
            "default": "capacity-norms-v1.0"
          },
          "calculation_version": {
            "type": "string",
            "title": "Calculation Version",
            "description": "Audit version of the mathematical calculation implementation.",
            "default": "calc-v1.0"
          }
        },
        "type": "object",
        "required": [
          "cc_land",
          "livelihood_multiplier"
        ],
        "title": "CapacityBreakdownDTO",
        "description": "Structured breakdown of independent resource capacity dimensions."
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Standardized machine-readable error code string."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable explanation of the error."
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "description": "Unique request tracing identifier."
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details",
            "description": "Contextual error details or validation breakdown."
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "description": "Standardized error details conforming to SETU-DRR API error envelope."
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail",
            "description": "Error details payload."
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "description": "Canonical API error response envelope."
      },
...
      "ExternalRecommendationItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "import_run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Import Run Id"
          },
          "habitation_id": {
            "type": "integer",
            "title": "Habitation Id"
          },
          "habitation_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Habitation Name"
          },
          "site_id": {
            "type": "integer",
            "title": "Site Id"
          },
          "external_habitation_key": {
            "type": "string",
            "title": "External Habitation Key"
          },
          "external_site_key": {
            "type": "string",
            "title": "External Site Key"
          },
          "origin_type": {
            "type": "string",
            "title": "Origin Type",
            "default": "external"
          },
          "decision_status": {
            "type": "string",
            "title": "Decision Status",
            "default": "recommendation"
          },
          "households": {
            "type": "integer",
            "title": "Households"
          },
          "tier": {
            "type": "string",
            "title": "Tier"
          },
          "priority_score": {
            "type": "number",
            "title": "Priority Score"
          },
          "distance_km": {
            "type": "number",
            "title": "Distance Km"
          },
          "site_suitability": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Suitability"
          },
          "site_cc_final": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Cc Final"
          },
          "site_binding": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Binding"
          },
          "has_group_split": {
            "type": "boolean",
            "title": "Has Group Split",
            "default": false
          },
          "rationale": {
            "additionalProperties": true,
            "type": "object",
            "title": "Rationale"
          },
          "screening_grade": {
            "type": "string",
            "title": "Screening Grade"
          },
          "screening_caveats": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Screening Caveats"
          },
          "source_pipeline": {
            "type": "string",
            "title": "Source Pipeline",
            "default": "external_gis_v1"
          },
          "pipeline_version": {
            "type": "string",
            "title": "Pipeline Version",
            "default": "v1.0"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "import_run_id",
          "habitation_id",
          "site_id",
          "external_habitation_key",
          "external_site_key",
          "households",
          "tier",
          "priority_score",
          "distance_km",
          "screening_grade",
          "created_at"
        ],
        "title": "ExternalRecommendationItem",
        "description": "External offline recommendation record from upstream partner pipeline."
      },
...
      "PaginatedResponse_CandidateSiteItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CandidateSiteItem"
            },
            "type": "array",
            "title": "Items",
            "description": "Page of records."
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total",
            "description": "Total matching records count."
          },
          "limit": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Limit",
            "description": "Page limit."
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset",
            "description": "Page offset."
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "description": "Whether subsequent records exist."
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset",
          "has_more"
        ],
        "title": "PaginatedResponse[CandidateSiteItem]"
      },
...
      "TenureType": {
        "type": "string",
        "enum": [
          "government_revenue",
          "private",
          "tenure_unverified"
        ],
        "title": "TenureType",
        "description": "Land tenure status for candidate relocation sites (PRD \u00a76.8, FR-7.3)."
      },
...
    }
  }
}